+++ src/main/java/org/jetbrains/car/PetrolCar.java
@@ -4,7 +4,7 @@ import org.jetbrains.location.Location;
 import org.jetbrains.util.Constants;
 
 public class PetrolCar extends Car {
-    public PetrolCar(Location location, int energyUsageRate) {
+    public PetrolCar(Location location, double energyUsageRate) {
         super(location, energyUsageRate);
         energyThreshold = Constants.FUEL_CAR_ENERGY_THRESHOLD;
     }
